Update Basic Cashier Calculator
You work for a local shipping company named East County Box Company. This company ships large boxes and crates across the country.

You have been tasked with making the cashiers' lives easier by creating a calculator that takes the size of the crate and the cost per cubic foot and determine the customer's cost.

Requirements:

You will ask the user for the width, length and height of the package
The program will then calculate the volume and display the volume
The volume will then be calculated against the standard rate of $2.50 per cubic foot and display that number
Because we live in California, the program then calculate the sales tax of .0775 per dollar and display that number
A total will be calculate and displayed. This creates the invoice for the transaction.