Create use functions in the calculator
This is an expansion on Mini Project 2. We will be incorporating functions into our program.

Requirements:

Modify your Mini Project 2 so that the section that determines the volume of the box is a function. The function will take in the height, length and width and it will return to the program the volume. You must demonstrate the use of function prototypes. 

The function will be called within your loops.