Create Delivery cost calculations
Now we are going to give delivery options:

This is the criteria:

Standard - (one to two weeks) - No change to the price
Fast Ground - (3 to 5 business days) - Extra $1.50 per cubic foot
Air - (1 to 2 business days) - Only applies to packages under 30 cubic feet - Extra $3.00 per cubic foot
Requirements:

Each package is to be assessed individually.
Use a single-key menu system to choose which delivery plan. Show the pricing schedule as I wrote it above.
For standard use 14 days; for fast use 5 days; and for air use 2 days. (Optional) On the final receipt, you will use the date function to estimate the delivery date (This is for Project 7). 
The extra delivery charge will be labeled as "Premium Delivery" on the receipt