Create main and secondary function
Create a program with your main and one supporting function that uses a reference between the main function and the support function. The main function initializes a variable with a value of 5, a call that sends the variable to the function, the function changes the variable to the value of 6 and the main function prints out the new value of the variable.